Let the ontology say what the catalog is
The catalog is now assembled from what the ontology entails, each act
bound by key to the code that carries it out, in key order. An act the
trees entail and nothing carries stops the program at start. Farming is
two acts, as the trees have it: clearing makes a field, farming harvests
one, so a field is a thing an agent has or lacks.

+++ core/observe/observe.go
@@ -231,7 +231,7 @@ func habits(w *world.World) (spread, mean, gated float64) {
 			}
 		}
 	}
-	mean /= n * action.Count
+	mean /= n * float64(action.Count)
 	if gatedN > 0 {
 		gated /= gatedN
 	}
@@ -250,7 +250,7 @@ func habits(w *world.World) (spread, mean, gated float64) {
 			spread += habit.Norm(d)
 		}
 	}
-	spread /= n * action.Count
+	spread /= n * float64(action.Count)
 	return spread, mean, gated
 }
